large change to equippable item icons
