+++ EquippableItemIcons/EquippableItemIcons.csproj
@@ -75,6 +75,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="API\ActivatedEquippableItem.cs" />
     <Compile Include="API\HudItemIcon.cs" />
     <Compile Include="API\Registries.cs" />
     <Compile Include="HudItemIconManager.cs" />
